+++ core/loaders/text.ts
@@ -1,4 +1,4 @@
-import { parseFrontMatter } from "./yaml.ts";
+import { extract, test } from "../../deps/front_matter.ts";
 import { read } from "../utils.ts";
 
 import type { Data } from "../../core.ts";
@@ -6,5 +6,13 @@ import type { Data } from "../../core.ts";
 /** Load a text file. Detect and parse the front matter */
 export default async function text(path: string): Promise<Data> {
   const content = await read(path, false);
-  return parseFrontMatter(content, path);
+
+  if (test(content)) {
+    const { attrs = {}, body } = extract<Data>(content);
+    attrs.content = body;
+
+    return attrs;
+  }
+
+  return { content };
 }